Always use Scripting Bridge to invoke GUI from CLI
When the CLI command is invoked without any arguments or with some refs,
the CLI command was using a low-level NSAppleEventDescriptor to
communicate with the main GUI process. That behavior was inconsistent
with the behavior for other command-line options, where communication is
done with high-level Scripting Bridge. Moreover, the GUI side was
failing to handle the NSAppleEventDescriptor and thus ignoring command
line arguments passed from CLI (cf. https://github.com/rowanj/gitx/issues/196).

This commit modifies the code so that CLI and GUI always use Scripting
Bridge to send and receive command-line arguments correctly.