Don't recreate graphQueue
Recreating the graph queue means that the new queue may start a
background operation before the old queue is still running another
operation. Multiple operations being executed at the same time seem to
be causing an occasional crash because they are not synchronized.

Note that calling cancelAllOperations for the old queue does not abort
an operation that has already started.