More work on working directory detection.
Here are the steps it goes through :
1. If `--git-dir` is provided and points to a repository, return that. If it's not a repo, inform the user and exit.
	Handles `gitx --git-dir=path` or `gitx --git-dir path`.

2. Loops through the rest of the arguments for something that resolves to a repository.
	Handles `gitx ..`, `gitx path/to/submodule`.

3. Try to find a repository starting from the current working directory.
	Handles `gitx`.

4. Inform user that we couldn't find a repository and exit.

I'm not yet sure about step 2, because it somewhat conflicts with the current behavior if there's a directory with the same name as a branch. You'd end up resolving `branchname` as a repository (either as itself if it is one, or as your CWD) and cause the GitX window to reopen instead of filtering on the `branch name` branch.